Create a plot utils module and add the basic orchard plot.
